xfstests: btrfs, test send's ability to punch holes, prealloc extents and send data size
This test verifies that after an incremental btrfs send the
replicated file has the same exact hole and data structure as in
the origin filesystem. This didn't use to be the case before the
send stream version 2 - holes were sent as write operations of 0
valued bytes instead of punching holes with the fallocate system
call, and pre-allocated extents were sent as well as write
operations of 0 valued bytes instead of intructions for the
receiver to use the fallocate system call.

It also checks that prealloc extents that lie beyond the file's
size are replicated by an incremental send.

Also update existing test btrfs/161 to verify total data size

[Howard: rebased on kernel v4.17-rc7, btrfs progs v4.16.1]
Signed-off-by: Howard McLauchlan <linux@hmclauchlan.com>
Signed-off-by: Filipe David Borba Manana <fdmanana@gmail.com>